HW - LangChain II/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.13
langchain-text-splitters>=0.0.1

# Embeddings and Vector Store  
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
huggingface-hub>=0.20.0

# Data processing (versión compatible con Python 3.14)
pandas>=2.2.0

# UI
streamlit>=1.29.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
colorama>=0.4.6
